Rename mutator file to be consistent
