+++ addon/model/model.ts
@@ -7,7 +7,7 @@ import ModelSelection from '@lblod/ember-rdfa-editor/model/model-selection';
 import ModelElement from "@lblod/ember-rdfa-editor/model/model-element";
 import SelectionReader from "@lblod/ember-rdfa-editor/model/readers/selection-reader";
 import SelectionWriter from "@lblod/ember-rdfa-editor/model/writers/selection-writer";
-import ModelMutator from "@lblod/ember-rdfa-editor/model/mutators/ModelMutator";
+import ModelMutator from "@lblod/ember-rdfa-editor/model/mutators/model-mutator";
 
 
 /**
